Separated the two classes in eos_spiner.hpp
